projeto/server_remoto/api/serializers.py
from rest_framework import serializers
from server_remoto import models

class PacienteSerializer(serializers.ModelSerializer):
    class Meta:
        model = models.Paciente
        fields = '__all__'
